+++ source/product.cpp
@@ -1,7 +1,7 @@
 #include "product.h"
 
 
-Product::Product(int id, const string &name, double price) : m_id(id), m_name(name), m_price(price) {}
+Product::Product(int id, const string &name, int price) : m_id(id), m_name(name), m_price(price) {}
 
 int Product::getId() const
 {
@@ -13,7 +13,7 @@ string Product::getName() const
     return m_name;
 }
 
-double Product::getPrice() const
+int Product::getPrice() const
 {
     return m_price;
 }
@@ -23,7 +23,7 @@ void Product::setName(const string &name)
     m_name = name;
 }
 
-void Product::setPrice(double price)
+void Product::setPrice(int price)
 {
     m_price = price;
 }
